Revert "[libcxx][pretty printers] Import gdb module in gdb feature check"
This reverts commit 964fc864e4ff2990cab1e2e00cf50d467fc301b9.

This was my mistake. When gdb can find its data directory it'll
import it automatically. If it can't (like when you're using a
version from a build folder) you need to give it the data
directory path.

We're safe to assume gdb is installed for testing purposes
so it'll import it for us.